requirements.txt
streamlit
google-generativeai
google-cloud-texttospeech
pillow
ffmpeg-python
python-dotenv
mutagen
mutagen
